+++ src/service-implementations/lf-repo-tree-node-service.ts
@@ -1,22 +1,33 @@
 import { LfRepoTreeNode } from '../helper-types/lf-repo-browser-types';
-import {
-  LfLocalizationService,
-  IconUtils,
-  PathUtils
-} from '@laserfiche/lf-js-utils';
+import { LfLocalizationService, IconUtils, PathUtils, StringUtils } from '@laserfiche/lf-js-utils';
 import { getFolderChildrenDefaultParameters } from '../utils/repo-client-utils.js';
-import { Entry, Document, Folder, Shortcut, RecordSeries, ODataValueContextOfIListOfEntry, EntryType, FindEntryResult } from '@laserfiche/lf-repository-api-client';
+import {
+  Entry,
+  Document,
+  Folder,
+  Shortcut,
+  RecordSeries,
+  ODataValueContextOfIListOfEntry,
+  EntryType,
+  FindEntryResult,
+} from '@laserfiche/lf-repository-api-client';
 import { IRepositoryApiClientEx } from '../helper-types/repository-api-ex.js';
-import { defaultSupportedColumnIds } from '../helper-types/lf-repo-browser-types';
-import { ColumnOrderBy, LfTreeNodePage, LfTreeNodeService, PropertyValue } from '@laserfiche/types-lf-ui-components';
-import { StringUtils } from '@laserfiche/lf-js-utils';
+import { allSupportedRepositoryColumnIds } from '../helper-types/lf-repo-browser-types';
+import {
+  ColumnOrderBy,
+  LfTreeNode,
+  LfTreeNodePage,
+  LfTreeNodeService,
+  PropertyValue,
+} from '@laserfiche/types-lf-ui-components';
+
 export const nodeAttrName_extension = 'extension';
 export const nodeAttrName_elecDocumentSize = 'elecDocumentSize';
 export const nodeAttrName_templateName = 'templateName';
 export const nodeAttrName_creationTime = 'creationTime';
+const rootFolderId: number = 1;
 
 export class LfRepoTreeNodeService implements LfTreeNodeService {
-
   /**
    * An array containing entryTypes (defined in '@laserfiche/lf-repository-api-client') for viewable entries
    * @example
@@ -51,9 +62,7 @@ export class LfRepoTreeNodeService implements LfTreeNodeService {
 
   private localizationService = new LfLocalizationService();
 
-  constructor(
-    private repoClient: IRepositoryApiClientEx,
-  ) {
+  constructor(private repoClient: IRepositoryApiClientEx) {
     if (!this.repoClient) {
       console.error('getParentNodeAsync() _selectedRepositoryClient was undefined');
       throw new Error(this.localizationService.getString('ERRORS.ENTRY_NOT_FOUND'));
@@ -87,7 +96,7 @@ export class LfRepoTreeNodeService implements LfTreeNodeService {
    */
   async getSupportedColumnIdsAsync(): Promise<string[]> {
     // TODO: add support for custom repository-specific columns
-    return defaultSupportedColumnIds;
+    return allSupportedRepositoryColumnIds;
   }
 
   /**
@@ -112,7 +121,6 @@ export class LfRepoTreeNodeService implements LfTreeNodeService {
    * ```
    */
   async getParentTreeNodeAsync(treeNode: LfRepoTreeNode): Promise<LfRepoTreeNode | undefined> {
-    const repoName: string = await this.repoClient.getCurrentRepoName();
     const repoId: string = await this.repoClient.getCurrentRepoId();
     if (treeNode.id === '1') {
       return undefined;
@@ -124,27 +132,31 @@ export class LfRepoTreeNodeService implements LfTreeNodeService {
     try {
       const parentPath = this.getParentPath(treeNode.path);
       if (parentPath) {
-        const parentEntry: FindEntryResult = await this.repoClient.entriesClient.getEntryByPath({ repoId, fullPath: parentPath });
+        const parentEntry: FindEntryResult = await this.repoClient.entriesClient.getEntryByPath({
+          repoId,
+          fullPath: parentPath,
+        });
         if (parentEntry.entry) {
           const foundParentEntry = parentEntry.entry;
           foundParentEntry.fullPath = parentPath;
-          return this.createLfRepoTreeNode(foundParentEntry, repoName);
-        }
-        else {
+          if (foundParentEntry.id === 1) {
+            const repoName = await this.repoClient.getCurrentRepoName();
+            return this.createRootFolderNode(repoName, foundParentEntry);
+          } else {
+            const parentNode = this.createNonRootLfRepoTreeNode(foundParentEntry);
+            return parentNode;
+          }
+        } else {
           throw new Error('Parent is not found');
         }
+      } else {
+        return await this.getRootTreeNodeAsync();
       }
-      else {
-        return this.getRootTreeNodeAsync();
-      }
-
-    }
-    catch (err: any) {
+    } catch (err: any) {
       if (err.errorCode === 9013) {
         const rootNode = await this.getRootTreeNodeAsync();
         return rootNode;
-      }
-      else {
+      } else {
         throw err;
       }
     }
@@ -192,20 +204,25 @@ export class LfRepoTreeNodeService implements LfTreeNodeService {
    *                                                    } // sorted by creationDate descending
    * ```
    **/
-  async getFolderChildrenAsync(folder: LfRepoTreeNode, nextPage?: string, orderBy?: ColumnOrderBy): Promise<LfTreeNodePage> {
-    const repoName: string = await this.repoClient.getCurrentRepoName();
+  async getFolderChildrenAsync(
+    folder: LfRepoTreeNode,
+    nextPage?: string,
+    orderBy?: ColumnOrderBy
+  ): Promise<LfTreeNodePage> {
     let listChildrenEntriesResponse: ODataValueContextOfIListOfEntry;
     if (!nextPage) {
       listChildrenEntriesResponse = await this.getFolderChildrenFirstPageAsync(folder, orderBy);
+    } else {
+      listChildrenEntriesResponse = await this.repoClient.entriesClient.getEntryListingNextLink({
+        nextLink: nextPage,
+        maxPageSize: 100,
+      });
     }
-    else {
-      listChildrenEntriesResponse = await this.repoClient.entriesClient.getEntryListingNextLink({ nextLink: nextPage, maxPageSize: 100 })
-    }
-    const dataMap = this.parseFolderChildrenResponse(folder, repoName, listChildrenEntriesResponse);
+    const dataMap = this.parseFolderChildrenResponse(folder, listChildrenEntriesResponse);
     const nextPageLink: string | undefined = listChildrenEntriesResponse.odataNextLink;
     return Promise.resolve({
       page: dataMap,
-      nextPage: nextPageLink
+      nextPage: nextPageLink,
     });
   }
 
@@ -221,111 +238,156 @@ export class LfRepoTreeNodeService implements LfTreeNodeService {
    *            - FolderInFolderInRoot
    *      - FolderInRoot2
    * const service = new LfRepoTreeNodeService(repoClient);
-   * service.getRootTreeNodeAsync() // returns the root node, with id of 1
+   * await service.getRootTreeNodeAsync() // returns the root node, with id of 1
    * ```
    */
   async getRootTreeNodeAsync(): Promise<LfRepoTreeNode> {
     const repoName: string = await this.repoClient.getCurrentRepoName();
     const repoId: string = await this.repoClient.getCurrentRepoId();
-    const rootFolderId: number = 1;
-    const rootEntry: Entry = await this.repoClient.entriesClient.getEntry(
-      {
-        repoId,
-        entryId: rootFolderId
-      }
-    );
-    return this.createLfRepoTreeNode(rootEntry, repoName);
-  }
-
-  /**
-   * Converts an Entry to the corresponding LfRepoTreeNode and returns it
-   * @param entry entry
-   * @param repoName Laserfiche repository name
-   * @param parent option parent LfRepoTreeNode
-   * @returns the corresponding LfRepoTreeNode
-   * @example
-   * ```ts
-   * const entry: Shortcut = createShortcut({
-   *  id: 14,
-   *  name: 'shortcutFolder',
-   *  fullPath: '\\shortcutFolder',
-   *  entryType: EntryType.Shortcut,
-   *  targetId: 20000,
-   *  targetType: EntryType.Folder
-   * });
-   * const service = new LfRepoTreeNodeService(repoClient);
-   * service.createLfRepoTreeNode(entry, 'repoName');
-   * // returns {
-   *  name: 'dummyShortcutFolder',
-   *  path: '\\dummyShortcutFolder',
-   *  id: '14',
-   *  icon: [
-   *   "https://lfxstatic.com/npm/@laserfiche/lf-resource-library@4/resources/icons/document-icons.svg#folder-20",
-   *   "https://lfxstatic.com/npm/@laserfiche/lf-resource-library@4/resources/icons/document-icons.svg#shortcut-overlay",
-   *  ],
-   *  isContainer: true,
-   *  isLeaf: false,
-   *  attributes: new Map<string, string>(),
-   *  entryType: EntryType.Shortcut,
-   *  targetId: 20000,
-   *  targetType: EntryType.Folder
-   * };
-   *  ```
-   */
-  createLfRepoTreeNode(entry: Entry, repoName: string, parent?: LfRepoTreeNode): LfRepoTreeNode {
-    let treeNode: LfRepoTreeNode | undefined;
-    if (!entry.entryType) {
-      throw new Error('entry type is undefined');
-    }
-    if (entry.entryType === EntryType.Folder) {
-      treeNode = this.createFolderNode(entry, repoName, parent);
-    } else if (entry.entryType === EntryType.Document) {
-      treeNode = this.createLeafNode(entry, parent);
-    } else if (entry.entryType === EntryType.Shortcut) {
-      const shortcut = entry as Shortcut;
-      if (shortcut.targetType === EntryType.Folder) {
-        treeNode = this.createFolderNode(shortcut, repoName, parent);
-      } else if (shortcut.targetType === EntryType.Document) {
-        treeNode = this.createLeafNode(shortcut, parent);
-      }
-      else if (shortcut.targetType === EntryType.RecordSeries) {
-        treeNode = this.createFolderNode(shortcut, repoName, parent);
-      }
-      else {
-        throw new Error('Unexpected shortcut targetType');
-      }
-    } else if (entry.entryType === EntryType.RecordSeries) {
-      treeNode = this.createFolderNode(entry, repoName, parent);
-    }
-    else {
-      throw new Error('Unsupported entry type');
-    }
-    this.setNodeProperties(treeNode, entry, parent);
+    const rootEntry: Entry = await this.repoClient.entriesClient.getEntry({
+      repoId,
+      entryId: rootFolderId,
+    });
+    const treeNode = this.createRootFolderNode(repoName, rootEntry);
     return treeNode;
   }
 
-  private isViewable(entry: Entry): boolean {
-    return (this.viewableEntryTypes?.includes(entry.entryType) && entry.entryType !== EntryType.Shortcut)
-      || (this.viewableEntryTypes?.includes(EntryType.Shortcut) && entry.entryType === EntryType.Shortcut && this.viewableEntryTypes?.includes((entry as Shortcut).targetType));
+  /**
+   * Returns the LfRepoTreeNode represented by the path
+   * @param identifier Unique identifier for the node, in this implementation the path of the node
+   * @returns The LfRepoTreeNode at the specified path, or throws if it does not exist
+   * @example
+   * ``` ts
+   * // suppose the repository is structured as below
+   * - root
+   *      - DocInRoot
+   *      - FolderInRoot1
+   *            - DocInFolderInRoot
+   *            - FolderInFolderInRoot
+   *      - FolderInRoot2
+   *            - ShortcutToFolderInRoot1
+   *                  - DocInFolderInRoot
+   *                  - FolderInFolderInRoot
+   * const rootNode = getTreeNodeByIdentifierAsync('\'); // returns the root node definition
+   * const nestedNode = getTreeNodeByIdentifierAsync('\FolderInRoot1\FolderInFolderInRoot'); // returns the definition for FolderInFolderInRoot
+   * const nonExistentNode = getTreeNodeByIdentifierAsync('\TestFolder'); // throws, will be caught by repository browser
+   * const nestedUnderShortcut = getTreeNodeByIdentifierAsync('\FolderInRoot2\ShortcutToFolderInRoot1\DocInFolderInRoot') // returns definition for DocInFolderInRoot, with path to shortcut rather than path to true location
+   * ```
+   */
+  async getTreeNodeByIdentifierAsync(identifier: string): Promise<LfTreeNode | undefined> {
+    const repoName: string = await this.repoClient.getCurrentRepoName();
+    const repoId: string = await this.repoClient.getCurrentRepoId();
+    const pathToNode: string = identifier;
+
+    const entryFound: FindEntryResult = await this.repoClient.entriesClient.getEntryByPath({
+      repoId,
+      fullPath: pathToNode,
+    });
+
+    if (entryFound.entry) {
+      const entryWithSpecifiedPath: Entry = entryFound.entry;
+      entryWithSpecifiedPath.fullPath = pathToNode;
+      let treeNode: LfTreeNode;
+      if (entryWithSpecifiedPath.id === 1) {
+        treeNode = this.createRootFolderNode(repoName, entryWithSpecifiedPath);
+      } else {
+        treeNode = this.createNonRootLfRepoTreeNode(entryWithSpecifiedPath);
+      }
+
+      return treeNode;
+    } else {
+      throw new Error(`Unable to get entry with path: ${pathToNode}`);
+    }
   }
 
-  private createLeafNode(entry: Entry, parent?: LfRepoTreeNode): LfRepoTreeNode {
-    const path = parent ? PathUtils.combinePaths(parent.path, entry.name) : entry.fullPath;
-    if (!path) {
-      throw new Error('entry fullPath is undefined');
+  private createRootFolderNode(repoName: string, rootEntry: Entry) {
+    const icon = this.getSingleIconForEntryType(EntryType.Folder);
+    const treeNode = this.createFolderNode(repoName, '\\', rootFolderId, EntryType.Folder, icon);
+    this.setColumnAttributesForEntry(rootEntry, treeNode);
+    return treeNode;
+  }
+
+  private createNonRootLfRepoTreeNode(entry: Entry, parent?: LfRepoTreeNode): LfRepoTreeNode {
+    let treeNode: LfRepoTreeNode | undefined;
+    let entryName: string | undefined = entry.name;
+    if (!entryName || entryName.length === 0) {
+      entryName = entry.id.toString();
+    }
+    let path: string = entry.fullPath;
+    if (!path && parent) {
+      path = this.getFullPath(parent, entryName);
     }
 
-    const leafNode: LfRepoTreeNode = {
-      name: entry.name!,
-      path: path,
-      id: entry.id!.toString(),
-      entryType: entry.entryType,
-      icon: [],
-      isContainer: false,
-      isLeaf: true,
-      attributes: new Map<string, PropertyValue>()
-    };
-    return leafNode;
+    if (!path) {
+      throw new Error(`Unable to determine path of entry: ${entry.id}`);
+    }
+
+    const targetEntryType: EntryType = (entry as Shortcut).targetType
+      ? (entry as Shortcut).targetType
+      : entry.entryType;
+
+    if (!targetEntryType) {
+      throw new Error(`Entry type is undefined for entry: ${entry.id}`);
+    }
+
+    const icon = this.getIconsForEntry(entry, parent);
+  
+    switch (targetEntryType) {
+      case EntryType.Folder:
+      case EntryType.RecordSeries:
+        treeNode = this.createFolderNode(entryName, path, entry.id, entry.entryType, icon);
+        break;
+      case EntryType.Document:
+        treeNode = this.createLeafNode(entryName, path, entry.id, entry.entryType, icon);
+        break;
+      default:
+        throw new Error(`Unsupported entry type for entry: ${entry.id}`);
+    }
+
+    this.setColumnAttributesForEntry(entry, treeNode);
+    this.setTargetIdAndTypeForShortcut(entry, treeNode);
+    return treeNode;
+  }
+
+  private getSingleIconForEntryType(entryType: EntryType, parent?: LfRepoTreeNode, extension?: string): string {
+    switch (entryType) {
+      case EntryType.Folder: {
+        const isParentShortcutRecordSeries =
+          parent?.entryType === EntryType.Shortcut && parent?.targetType === EntryType.RecordSeries;
+        const isParentRecordSeries = parent?.entryType === EntryType.RecordSeries;
+        if (isParentRecordSeries || isParentShortcutRecordSeries) {
+          return IconUtils.getDocumentIconUrlFromIconId('recordfolder-20');
+        } else {
+          return IconUtils.getDocumentIconUrlFromIconId('folder-20');
+        }
+      }
+      case EntryType.RecordSeries: {
+        return IconUtils.getDocumentIconUrlFromIconId('recordseries-20');
+      }
+      case EntryType.Document: {
+        if (extension) {
+          const iconId = IconUtils.getDocumentIconIdFromExtension(extension);
+          return IconUtils.getDocumentIconUrlFromIconId(iconId);
+        } else {
+          return IconUtils.getDocumentIconUrlFromIconId('document-20');
+        }
+      }
+      case EntryType.Shortcut: {
+        return IconUtils.getDocumentIconUrlFromIconId('shortcut-overlay');
+      }
+      default: {
+        throw new Error('unsupported entry type');
+      }
+    }
+  }
+
+  private isViewable(entry: Entry): boolean {
+    return (
+      (this.viewableEntryTypes?.includes(entry.entryType) && entry.entryType !== EntryType.Shortcut) ||
+      (this.viewableEntryTypes?.includes(EntryType.Shortcut) &&
+        entry.entryType === EntryType.Shortcut &&
+        this.viewableEntryTypes?.includes((entry as Shortcut).targetType))
+    );
   }
 
   private valueToPropertyValue(value: string | Date | number | undefined, columnId: string | undefined): PropertyValue {
@@ -335,7 +397,14 @@ export class LfRepoTreeNodeService implements LfTreeNodeService {
       case 'creationTime':
       case 'lastModifiedTime': {
         const date = new Date(value);
-        displayValue = new Intl.DateTimeFormat('en-US', { year: 'numeric', month: 'numeric', day: 'numeric', hour: 'numeric', minute: 'numeric', second: 'numeric' }).format(date);
+        displayValue = new Intl.DateTimeFormat('en-US', {
+          year: 'numeric',
+          month: 'numeric',
+          day: 'numeric',
+          hour: 'numeric',
+          minute: 'numeric',
+          second: 'numeric',
+        }).format(date);
         break;
       }
       case 'elecDocumentSize': {
@@ -350,58 +419,7 @@ export class LfRepoTreeNodeService implements LfTreeNodeService {
     return { value: value, displayValue: displayValue } as PropertyValue;
   }
 
-  private setNodeProperties(node: LfRepoTreeNode, entry: Entry, parent?: LfRepoTreeNode): void {
-    this.setAttributesForEntry(entry, node);
-    if (entry.entryType === EntryType.Document) {
-      const document = entry as Document;
-      if (document.extension) {
-        const iconId = IconUtils.getDocumentIconIdFromExtension(document.extension);
-        node.icon = IconUtils.getDocumentIconUrlFromIconId(iconId);
-      }
-      else {
-        node.icon = IconUtils.getDocumentIconUrlFromIconId('document-20');
-      }
-    }
-    else if (entry.entryType === EntryType.Folder) {
-      if (parent?.entryType === EntryType.RecordSeries ||
-        (parent?.entryType === EntryType.Shortcut && parent?.targetType === EntryType.RecordSeries)) {
-        node.icon = IconUtils.getDocumentIconUrlFromIconId('recordfolder-20');
-      }
-      else {
-        node.icon = IconUtils.getDocumentIconUrlFromIconId('folder-20');
-      }
-    }
-    else if (entry.entryType === EntryType.RecordSeries) {
-      node.icon = IconUtils.getDocumentIconUrlFromIconId('recordseries-20');
-    }
-    else if (entry.entryType === EntryType.Shortcut) {
-      const shortcut = entry as Shortcut;
-      // TODO: shortcut doesn't have elecDocumentSize
-      node.targetType = shortcut.targetType;
-      node.targetId = shortcut.targetId;
-      if (shortcut.targetType === EntryType.Document) {
-        const iconId = shortcut.extension ? IconUtils.getDocumentIconIdFromExtension(shortcut.extension) : undefined;
-        const iconUrl = IconUtils.getDocumentIconUrlFromIconId(iconId ?? 'document-20');
-        const shortcutUrl = IconUtils.getDocumentIconUrlFromIconId('shortcut-overlay');
-        node.icon = [iconUrl, shortcutUrl];
-      } else if (shortcut.targetType === EntryType.Folder) {
-        const iconUrl = IconUtils.getDocumentIconUrlFromIconId('folder-20');
-        const shortcutUrl = IconUtils.getDocumentIconUrlFromIconId('shortcut-overlay');
-        node.icon = [iconUrl, shortcutUrl];
-      } else if (shortcut.targetType === EntryType.RecordSeries) {
-        const iconUrl = IconUtils.getDocumentIconUrlFromIconId('recordseries-20');
-        const shortcutUrl = IconUtils.getDocumentIconUrlFromIconId('shortcut-overlay');
-        node.icon = [iconUrl, shortcutUrl];
-      } else {
-        throw new Error('unsupported shortcut targetType');
-      }
-    }
-    else {
-      throw new Error('unsupported entryType');
-    }
-  }
-
-  private setAttributesForEntry(entry: Document | Folder | Shortcut | RecordSeries, node: LfRepoTreeNode) {
+  private setColumnAttributesForEntry(entry: Document | Folder | Shortcut | RecordSeries, node: LfRepoTreeNode) {
     if (this.columnIds) {
       for (const columnId of this.columnIds) {
         node.attributes.set(columnId, this.valueToPropertyValue(entry[columnId], columnId));
@@ -409,76 +427,106 @@ export class LfRepoTreeNodeService implements LfTreeNodeService {
     }
   }
 
-  private createFolderNode(entry: Entry, repoName: string, parent?: LfRepoTreeNode): LfRepoTreeNode {
-    if (!entry.id) {
-      throw new Error('entryId is undefined');
+  private setTargetIdAndTypeForShortcut(entry: Entry, node: LfRepoTreeNode) {
+    if (entry.entryType === EntryType.Shortcut) {
+      const shortcut = entry as Shortcut;
+      node.targetId = shortcut.targetId;
+      node.targetType = shortcut.targetType;
     }
-    if (parent && !parent.path) {
-      throw new Error('path is undefined on parent entry');
-    }
-    let entryName: string | undefined = entry.name;
-    let path: string;
-    if (entry.id === 1) {
-      // rootNode does not have a name, default into repoName
-      path = '\\';
-      if (!entryName || entryName.length === 0) {
-        entryName = repoName;
-      }
-    }
-    else {
-      if (!entryName || entryName.length === 0) {
-        entryName = entry.id.toString();
-      }
-      path = parent ? PathUtils.combinePaths(parent.path, entryName) : entry.fullPath;
-      if (!path) {
-        throw new Error('fullPath is undefined');
-      }
-      else if (path.endsWith('\\') && path.length > 1) {
-        path = path.slice(0, -1);
-      }
-    }
+  }
 
+  private getIconsForEntry(entry: Entry, parent?: LfRepoTreeNode) {
+    const entryType = entry.entryType;
+    const extension = (entry as Document).extension;
+    let icon: string | string[] = this.getSingleIconForEntryType(entryType, parent, extension);
+    if (entry.entryType === EntryType.Shortcut) {
+      const shortcutIcon = icon;
+      const targetIcon = this.getSingleIconForEntryType((entry as Shortcut).targetType, parent, extension);
+      icon = [targetIcon, shortcutIcon];
+    }
+    return icon;
+  }
+
+  private getFullPath(parent: LfRepoTreeNode, entryName: string) {
+    let path = PathUtils.combinePaths(parent.path, entryName);
+    if (path.endsWith('\\') && path.length > 1) {
+      path = path.slice(0, -1);
+    }
+    return path;
+  }
+
+  private createFolderNode(
+    entryName: string,
+    path: string,
+    id: number,
+    entryType: EntryType,
+    icon: string | string[]
+  ): LfRepoTreeNode {
     const folderNode: LfRepoTreeNode = {
       name: entryName,
-      path: path,
-      id: entry.id!.toString(),
-      entryType: entry.entryType,
-      icon: [],
+      path,
+      id: id.toString(),
+      entryType,
+      icon,
       isContainer: true,
       isLeaf: false,
-      attributes: new Map<string, PropertyValue>()
+      attributes: new Map<string, PropertyValue>(),
     };
     return folderNode;
   }
 
-  private async getFolderChildrenFirstPageAsync(folder: LfRepoTreeNode, orderBy?: ColumnOrderBy): Promise<ODataValueContextOfIListOfEntry> {
+  private createLeafNode(
+    entryName: string,
+    path: string,
+    id: number,
+    entryType: EntryType,
+    icon: string | string[]
+  ): LfRepoTreeNode {
+    const leafNode: LfRepoTreeNode = {
+      name: entryName,
+      path,
+      id: id.toString(),
+      entryType,
+      icon,
+      isContainer: false,
+      isLeaf: true,
+      attributes: new Map<string, PropertyValue>(),
+    };
+    return leafNode;
+  }
+
+  private async getFolderChildrenFirstPageAsync(
+    folder: LfRepoTreeNode,
+    orderBy?: ColumnOrderBy
+  ): Promise<ODataValueContextOfIListOfEntry> {
     let entryId: number;
     if (folder.targetId) {
       entryId = folder.targetId;
-    }
-    else {
+    } else {
       entryId = parseInt(folder.id, 10);
     }
     const repoId: string = await this.repoClient.getCurrentRepoId();
-    if (orderBy && !defaultSupportedColumnIds.includes(orderBy.columnId)) {
+    if (orderBy && !allSupportedRepositoryColumnIds.includes(orderBy.columnId)) {
       orderBy = undefined;
       console.error(`Cannot order by unsupported column: ${orderBy.columnId}`);
     }
 
     const requestParameters = getFolderChildrenDefaultParameters(repoId, entryId, this.columnIds, orderBy);
-    const listChildrenEntriesResponse: ODataValueContextOfIListOfEntry = await this.repoClient.entriesClient.getEntryListing(
-      requestParameters
-    );
+    const listChildrenEntriesResponse: ODataValueContextOfIListOfEntry =
+      await this.repoClient.entriesClient.getEntryListing(requestParameters);
     return listChildrenEntriesResponse;
   }
 
-  private parseFolderChildrenResponse(parent: LfRepoTreeNode, repoName: string, listChildrenEntriesResponse: ODataValueContextOfIListOfEntry): LfRepoTreeNode[] {
+  private parseFolderChildrenResponse(
+    parent: LfRepoTreeNode,
+    listChildrenEntriesResponse: ODataValueContextOfIListOfEntry
+  ): LfRepoTreeNode[] {
     const dataMap: LfRepoTreeNode[] = [];
     const childrenEntries: Entry[] | undefined = listChildrenEntriesResponse.value;
     if (childrenEntries) {
       for (const childEntry of childrenEntries) {
         if (this.isViewable(childEntry)) {
-          const childNode: LfRepoTreeNode = this.createLfRepoTreeNode(childEntry, repoName, parent);
+          const childNode: LfRepoTreeNode = this.createNonRootLfRepoTreeNode(childEntry, parent);
           dataMap.push(childNode);
         }
       }
@@ -495,4 +543,3 @@ export class LfRepoTreeNodeService implements LfTreeNodeService {
     return paths.join('\\');
   }
 }
-